handle show list category, product, variant with metadata

+++ src/services/Admin/categoryService.js
@@ -1,4 +1,3 @@
-import { ObjectId } from "mongodb";
 import { client } from "../../config/mongodb.js";
 import categoryModel from "../../models/categoryModel.js";
 import Category from "../../models/categoryModel.js"
@@ -8,14 +7,54 @@ const getAll = async () => {
   return await Category.getAll();
 }
 
+const getAllWithMetadata = async (query = {}) => {
+  let conditions = []
+  if (query && query.active) {
+    switch (query.active) {
+      case 1:
+        conditions.push(
+          { is_active: true }
+        )
+        break;
+      case 0:
+        conditions.push(
+          { is_active: false }
+        )
+    }
+  }
+  if (query.search) {
+    conditions.push({
+      $or: [
+        {
+          name: {
+            $regex: query.search.trim(), $options: "i"
+          }
+        }
+      ]
+    })
+  }
+
+  conditions = conditions.length > 0 ? { $and: conditions } : {}
+  const categories = await categoryModel.getAll({ conditions, query })
+  const total = await categoryModel.countAll();
+  const totalFiltered = await categoryModel.countFiltered(query)
+  return {
+    items: categories,
+    meta: {
+      total,
+      totalFiltered
+    }
+  }
+}
+
 const create = async (data) => {
-  
+
   const exist = await Category.findBy({ name: data?.name });
 
   if (exist) {
     throw new ErrorCustom('Tên danh mục đã tồn tại!', 400)
   }
-  
+
   if (data.parent_id) {
     const existParentCategory = await Category.findById(data.parent_id);
     if (!existParentCategory) {
@@ -52,8 +91,8 @@ const update = async (id, data) => {
 }
 
 const findById = async (id) => {
-  const result =  await categoryModel.findById(id);
-  if(!result) throw new ErrorCustom("Không tìm thấy danh mục sản phẩm");
+  const result = await categoryModel.findById(id);
+  if (!result) throw new ErrorCustom("Không tìm thấy danh mục sản phẩm");
   return result
 }
 
@@ -62,10 +101,13 @@ const destroy = async (id) => {
   const session = await client.startSession()
   try {
     session.startTransaction();
-    // Tìm xem category có tồn tại không
     const category = await Category.findById(id);
     if (!category) {
       throw new ErrorCustom('Danh mục không tồn tại!', 404);
+    }
+
+    if(category?.type == 'not_delete'){
+      throw new ErrorCustom('Danh mục này không thể xoá!', 419)
     }
     // Xoá các danh mục con
     await Category.deleteChildrenByIdParent(id, session);
@@ -81,7 +123,7 @@ const destroy = async (id) => {
 }
 
 const getChildrenCategory = async (parentId) => {
-  
+
   const parentCategory = await Category.findById(parentId);
   if (!parentCategory) {
     throw new ErrorCustom('Danh mục cha không tồn tại!', 404)
@@ -100,6 +142,7 @@ const getParentCategory = async (idParent) => {
 
 export default {
   getAll,
+  getAllWithMetadata,
   create,
   update,
   findById,
